S15.e.6: extract operator.ts spawn dialog (1279 → 1027 LOC)
Eighth sub-slice of the S15.e operator.ts decomposition train.
Lifts the heaviest remaining closure-captured block — the n-key
spawn-agent dialog — into a dedicated dialogs module.

cli/src/commands/operator/dialogs/spawn.ts (~295 LOC):
  openSpawnDialog(ctx: SpawnDialogContext) containing the original
  draw / close / startEdit / launch / onKey nested functions.
  Modal flag passes through setDialogOpen(open) so the caller still
  owns dialogOpen.

A thin screen.key(['n'], ...) wrapper preserves the keymap binding
in operator.ts.

Behavioral change: NONE. Body byte-identical (267 LOC excised) modulo
the dialogOpen = ... → setDialogOpen(...) callback rewrite.

Cumulative S15.e delta:
  pre-S15.e: 2894
  e.1-5c:    1279  (-1615)
  e.6:       1027   (-252)
  Cumulative: -1867. Connect/delete modals (~230 LOC) remaining for cap.

Verification:
- tsc --noEmit clean / lint 22 (+2 any-type for blessed.Screen) / build clean
- npm test: 454 pass / 2 skipped pre-existing

Audit: docs/security-audits/2026-04-29-phase2-hotspot-operator-cli-e6.md
